Implemented a singleton class and applied another way to calculate face distance.(Refer to the details)
[Reason of Merge]

1. to apply singleton class
whole visitors data should be used across all the methods(list, create, destroy, partial_update, etc.), and the corresponding data should be consistent among those methods.

2. to make the procedure clear to developers
there are three procedures(below) and by getting them together, the procedures can be clearer
- updating queryset(to invoke SQL query in DB)
- updating serializer(a method to convert from django class to json format)
- updating visitors data(uuid and face encoding dictionary)

3. to apply another way to calculate face distance
previous methods used a way to output a list which has all the distance with a current visitor. But this has a problem if there're more than 1 data which has closer distance(than a threshold) with the current visitor.(the current visitor should match only one data in DB). So improved method applies better way to calculate distance, which takes only the closest distance, and if there aren't any data which is closest enough, the server recognizes the current visitor as a new data to be inserted in DB as a new record.

[history of commits]

* created an environment of docker-compose.

* solved some errors in the issues, and testing execution environments

* edited run.py

* tested execution environment and editted recirements.txt in the top directory

* solved some errors, and modified some texts

* modified several files

* added empty file

* modified frontenddockerfile

* added a command to create images directory for temporarily saving images

* modified frontend dockerfile

* implemented a singleton class for VisitorsData class

* Refactored to create a singleton class for visitors_data and changed a way to calculate face_distance of visitors